contract: refactor seal-specific types into seals module
